+++ include/libfly/potential/EAM/eam.hpp
@@ -84,12 +84,14 @@ namespace fly::potential {
      *
      * Assumes the neighbour list are ready, force on frozen atoms will be zero.
      *
-     * @param out Per-atom TypeID's and Frozen properties
-     * @param in Per-atom Potential gradient written to this.
+     * @param in Per-atom TypeID's and Frozen properties
+     * @param out Potential gradient written to this.
      * @param nl Neighbour list (in ready state i.e. neigh::List::update() or neigh::List::rebuild() called).
      * @param threads Number of openMP threads to use.
      */
-    auto gradient(system::SoA<PotentialGradient&> in, system::SoA<TypeID const&, Frozen const&> out, neigh::List const& nl,
+    auto gradient(system::SoA<PotentialGradient&> out,
+                  system::SoA<TypeID const&, Frozen const&> in,
+                  neigh::List const& nl,
                   int threads = 1) -> void;
 
     /**