+++ include/libfly/utility/spline.hpp
@@ -15,6 +15,7 @@
 
 // You should have received a copy of the GNU General Public License along with openFLY. If not, see <https://www.gnu.org/licenses/>.
 
+#include <algorithm>
 #include <string>
 #include <vector>